\name{insertLine}
\alias{insertLine}
\title{Build and insert line}
\usage{
insertLine(dashnb)
}
\arguments{
\item{dashnb}{Number of "-" to be inserted after "#"}
}
\description{
This is the more general function that builds and then print  the commented line
via rstudioapi. The argument dashnb specifies the number of "-" to be printed
after the "#".
}
\examples{
\dontrun{insertDashLine()}
}
